+++ BertheVarioPlatformIO/src/GlobalVar/CGlobalVar.h
@@ -43,7 +43,7 @@ public :
 
     // pour affichage
     float m_VitVertMS ;     ///< vitesse verticale
-    int   m_DureeVolMin ;   ///< duree vol en minutes
+    float m_DureeVolMin ;   ///< duree vol en minutes
 
     // pour elaboration finesse
     CTerrainsConnu m_TerrainArr ;     ///< fichier de terrains connus pour elaboration finesse
